Add popups to confirm deletion and show loading message

+++ Ionic/conFusion/www/js/controllers.js
@@ -230,20 +230,32 @@ angular.module('conFusion.controllers', [])
             }])
 
         .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory',
-            'baseURL', '$ionicListDelegate',
-            function($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+            'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout',
+            function($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate,
+                    $ionicPopup, $ionicLoading, $timeout) {
                 $scope.baseURL = baseURL;
                 $scope.shouldShowDelete = false;
-                $scope.message = "Loading ...";
+
+                $ionicLoading.show({
+                    template: '<ion-spinner></ion-spinner> Loading...'
+                });
 
                 $scope.favorites = favoriteFactory.getFavorites();
 
                 menuFactory.getDishes().query(
                     function(response) {
                         $scope.dishes = response;
+                        // Simulate delay
+                        $timeout(function() {
+                          $ionicLoading.hide();
+                        }, 1000);
                     },
                     function(response) {
                         $scope.message = "Error: " + response.status + " " + response.statusText;
+                        // Simulate delay
+                        $timeout(function() {
+                          $ionicLoading.hide();
+                        }, 1000);
                 });
                 console.log($scope.dishes, $scope.favorites);
 
@@ -253,7 +265,21 @@ angular.module('conFusion.controllers', [])
                 };
 
                 $scope.deleteFavorite = function (index)    {
-                    favoriteFactory.deleteFromFavorites(index);
+
+                    var confirmPopup = $ionicPopup.confirm({
+                        title: 'Confirm Delete',
+                        template: 'Are you sure you want to delete this item?'
+                    });
+
+                    confirmPopup.then(function (res)    {
+                        if (res)    {
+                            console.log('Ok to delete');
+                            favoriteFactory.deleteFromFavorites(index);
+                        }   else {
+                            console.log('Canceled delete');
+                        }
+                    });
+
                     $scope.shouldShowDelete = false;
                 };
